# storage/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Circle of life Api Documentation",
        "description": "Implementation of Circle of life",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.cpoints.net",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/habit": {
            "get": {
                "tags": [
                    "Habits"
                ],
                "summary": "Get list of habits",
                "description": "Returns list of habits",
                "operationId": "4-gethabitList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Habits"
                ],
                "summary": "Crete habit  ",
                "operationId": "create-habit-data",
                "requestBody": {
                    "description": "Pass habit data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "section_id",
                                    "points",
                                    "date_type",
                                    "repetition_type",
                                    "repetition_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "String"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "test.png"
                                    },
                                    "points": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "section_id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "date_type": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "repetition_type": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "repetition_number": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/habit/{id}": {
            "get": {
                "tags": [
                    "Habits"
                ],
                "summary": "Get habit",
                "description": "Returns habit information",
                "operationId": "5-get-habit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Habits"
                ],
                "summary": "habit Update",
                "operationId": "6-Update-habit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update User data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "move_type": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Habits"
                ],
                "summary": "delete habit",
                "operationId": "7-remove-habit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/update-habit/{id}": {
            "put": {
                "tags": [
                    "Habits"
                ],
                "summary": "habit Update",
                "operationId": "6-Update-full-habit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update User data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "String"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "test.png"
                                    },
                                    "points": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "section_id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "date_type": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "repetition_type": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "repetition_number": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}